Use the platform linker by default, which will just be a call to "cc" (the Win solution will be done later)

+++ src/compiler_tests/benchmark.c
@@ -9,7 +9,7 @@ static int begin = 0;
 
 void bench_begin(void)
 {
-	begin = clock();
+	begin = (int)clock();
 }
 double bench_mark(void)
 {
